pyproject.toml
[project]
name = "langchain-code-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autopep8>=2.3.2",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "langchain-groq>=0.3.6",
    "langchain-sandbox>=0.0.6",
    "langgraph>=0.6.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
